Refactor the WarcraftLogs links to be more reusable
