+++ .env
@@ -6,6 +6,7 @@ REACT_APP_SERVER_BASE=/
 REACT_APP_API_BASE=api/
 REACT_APP_WOW_DATABASE=wowdb
 REACT_APP_RAVEN_DSN=https://95bf5a7af57f4a57af1f1be4d0706a91@sentry.io/232829
+REACT_APP_WARCRAFT_LOGS_DOMAIN=https://www.warcraftlogs.com/
 
 # Game info
 REACT_APP_CURRENT_GAME_PATCH=8.0.0
